LegacyBios: Update EFI_COMPATIBILITY16_TABLE to match 0.98 CSM spec update
The addition of UmaAddress/UmaSize fields allows the CSM to have
writable memory between the top of the option ROMs and the start of
its read-only code segment.

The HiPermanentMemoryAddress/HiPermanentMemorySize fields added in the
new spec are also added to the table, but not yet supported.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: David Woodhouse <David.Woodhouse@intel.com>